Lesson6 task6 в процессе
