--- Ehrlich_Richard_Lesson6/hobby.csv
@@ -1,2 +1,0 @@
-скалолазание, охота,
-горные лыжи
