+++ Ehrlich_Richard_Lesson6/task1 and task2/task2.py
