--- Ehrlich_Richard_Lesson6/users.csv
@@ -1,2 +1,0 @@
-Иванов Иван Иванович,
-Петров Петр Петрович
